Use commit hash (not tag) for `action-read-yaml` action version specification (#325)
Currently we are referencing a tag when specifying the version of the `action-read-yaml` action, which is potentially mutable.

[It was requested](https://github.com/hazelcast/hazelcast-mono/pull/1181#discussion_r1537109860) that the commit SHA be used instead, as per the [GitHub actions security hardening guide](https://github.com/hazelcast/hazelcast-mono/pull/1181#discussion_r1537109860).

Updated the version specification to the commit hash of the {currently used version](https://github.com/pietrobolcato/action-read-yaml/releases/tag/1.1.0) - i.e. no changes to the actual version of the action being used.